@@ -3,11 +3,13 @@
   "version": "0.1.0",
   "description": "gRPC-Web simple example",
   "devDependencies": {
+    "@babel/core": "^7.7.7",
     "@grpc/proto-loader": "^0.3.0",
+    "es3ify-webpack-plugin": "^0.1.0",
     "google-protobuf": "^3.6.1",
     "grpc": "^1.15.0",
     "grpc-web": "^1.0.0",
-    "webpack": "^4.16.5",
+    "webpack": "^4.41.4",
     "webpack-cli": "^3.1.0"
   }
 }
